# optimize eager test speed by relax autotest in some key eager tests