--- a/python/oneflow/test/modules/test_activation.py
+++ b/python/oneflow/test/modules/test_activation.py
@@ -29,7 +29,7 @@ import oneflow.unittest
 
 @flow.unittest.skip_unless_1n1d()
 class TestReLUModule(flow.unittest.TestCase):
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_relu_module_with_random_data(test_case):
         m = torch.nn.ReLU()
         m.train(random())
@@ -39,7 +39,7 @@ class TestReLUModule(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest()
+    @autotest(n=5)
     def test_relu_module_with_0dim_data(test_case):
         m = torch.nn.ReLU()
         m.train(random())
@@ -49,7 +49,7 @@ class TestReLUModule(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest(auto_backward=False, check_graph=True)
+    @autotest(n=5, auto_backward=False)
     def test_relu_module_with_0_size_data(test_case):
         m = torch.nn.ReLU()
         m.train(random())
@@ -67,7 +67,7 @@ class TestReLUModule(flow.unittest.TestCase):
 
 @flow.unittest.skip_unless_1n1d()
 class TestReLU6Module(flow.unittest.TestCase):
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_relu6_module_with_random_data(test_case):
         m = torch.nn.ReLU6()
         m.train(random())
@@ -77,7 +77,7 @@ class TestReLU6Module(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest()
+    @autotest(n=5)
     def test_relu6_module_with_0dim_data(test_case):
         m = torch.nn.ReLU6()
         m.train(random())
@@ -87,7 +87,7 @@ class TestReLU6Module(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest(auto_backward=False, check_graph=True)
+    @autotest(n=5, auto_backward=False)
     def test_relu6_module_with_0_size_data(test_case):
         m = torch.nn.ReLU6()
         m.train(random())
@@ -105,7 +105,7 @@ class TestReLU6Module(flow.unittest.TestCase):
 
 @flow.unittest.skip_unless_1n1d()
 class TestTanh(flow.unittest.TestCase):
-    @autotest()
+    @autotest(n=5)
     def test_tanh_module_with_random_data(test_case):
         m = torch.nn.Tanh()
         m.train(random())
@@ -115,7 +115,7 @@ class TestTanh(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest()
+    @autotest(n=5)
     def test_tanh_module_with_0dim_data(test_case):
         m = torch.nn.Tanh()
         m.train(random())
@@ -125,7 +125,7 @@ class TestTanh(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest(auto_backward=False, check_graph=True)
+    @autotest(n=5, auto_backward=False)
     def test_tanh_module_with_0_size_data(test_case):
         m = torch.nn.Tanh()
         m.train(random())
@@ -135,21 +135,21 @@ class TestTanh(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_flow_tanh_with_random_data(test_case):
         device = random_device()
         x = random_tensor().to(device)
         y = torch.tanh(x)
         return y
 
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_flow_tanh_with_0dim_data(test_case):
         device = random_device()
         x = random_tensor(ndim=0).to(device)
         y = torch.tanh(x)
         return y
 
-    @autotest(auto_backward=False, check_graph=True)
+    @autotest(n=5, auto_backward=False)
     def test_flow_tanh_with_0_size_data(test_case):
         device = random_device()
         x = random_tensor(4, 2, 3, 0, 3).to(device)
@@ -164,7 +164,7 @@ class TestTanh(flow.unittest.TestCase):
 
 @flow.unittest.skip_unless_1n1d()
 class TestELUModule(flow.unittest.TestCase):
-    @autotest()
+    @autotest(n=5)
     def test_elu_module_with_random_data(test_case):
         m = torch.nn.ELU(alpha=random() | nothing())
         m.train(random())
@@ -174,7 +174,7 @@ class TestELUModule(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest()
+    @autotest(n=5)
     def test_elu_module_with_0dim_data(test_case):
         m = torch.nn.ELU(alpha=random() | nothing())
         m.train(random())
@@ -184,7 +184,7 @@ class TestELUModule(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest(auto_backward=False, check_graph=True)
+    @autotest(n=5, auto_backward=False)
     def test_elu_module_with_0_size_data(test_case):
         m = torch.nn.ELU(alpha=random() | nothing())
         m.train(random())
@@ -197,7 +197,7 @@ class TestELUModule(flow.unittest.TestCase):
 
 @flow.unittest.skip_unless_1n1d()
 class TestCELUModule(flow.unittest.TestCase):
-    @autotest()
+    @autotest(n=5)
     def test_celu_module_with_random_data(test_case):
         m = torch.nn.CELU(alpha=random() | nothing())
         m.train(random())
@@ -207,7 +207,7 @@ class TestCELUModule(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest()
+    @autotest(n=5)
     def test_celu_module_with_0dim_data(test_case):
         m = torch.nn.CELU(alpha=random() | nothing())
         m.train(random())
@@ -217,7 +217,7 @@ class TestCELUModule(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest(auto_backward=False, check_graph=True)
+    @autotest(n=5, auto_backward=False)
     def test_celu_module_with_0_size_data(test_case):
         m = torch.nn.CELU(alpha=random() | nothing())
         m.train(random())
@@ -240,7 +240,7 @@ class TestCELUModule(flow.unittest.TestCase):
 
 @flow.unittest.skip_unless_1n1d()
 class TestGelu(flow.unittest.TestCase):
-    @autotest()
+    @autotest(n=5)
     def test_gelu_module_with_random_data(test_case):
         m = torch.nn.GELU()
         m.train(random())
@@ -250,7 +250,7 @@ class TestGelu(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest()
+    @autotest(n=5)
     def test_gelu_module_with_0dim_data(test_case):
         m = torch.nn.GELU()
         m.train(random())
@@ -263,7 +263,7 @@ class TestGelu(flow.unittest.TestCase):
 
 @flow.unittest.skip_unless_1n1d()
 class TestSigmoidModule(flow.unittest.TestCase):
-    @autotest()
+    @autotest(n=5)
     def test_sigmoid_module_with_random_data(test_case):
         m = torch.nn.Sigmoid()
         m.train(random())
@@ -273,7 +273,7 @@ class TestSigmoidModule(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest()
+    @autotest(n=5)
     def test_sigmoid_module_with_0dim_data(test_case):
         m = torch.nn.Sigmoid()
         m.train(random())
@@ -283,28 +283,28 @@ class TestSigmoidModule(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_sigmoid_flow_with_random_data(test_case):
         device = random_device()
         x = random_tensor().to(device)
         y = torch.sigmoid(x)
         return y
 
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_sigmoid_flow_with_0dim_data(test_case):
         device = random_device()
         x = random_tensor(ndim=0).to(device)
         y = torch.sigmoid(x)
         return y
 
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_sigmoid_tensor_with_random_data(test_case):
         device = random_device()
         x = random_tensor().to(device)
         y = x.sigmoid()
         return y
 
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_sigmoid_tensor_with_0dim_data(test_case):
         device = random_device()
         x = random_tensor(ndim=0).to(device)
@@ -357,7 +357,7 @@ class TestHardsigmoidModule(flow.unittest.TestCase):
         for arg in GenArgList(arg_dict):
             test_hardsigmoid_inplace_impl(test_case, *arg)
 
-    @autotest()
+    @autotest(n=5)
     def test_hardsigmoid_module_with_random_data(test_case):
         m = torch.nn.Hardsigmoid()
         m.train(random())
@@ -367,7 +367,7 @@ class TestHardsigmoidModule(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest()
+    @autotest(n=5)
     def test_hardsigmoid_module_with_0dim_data(test_case):
         m = torch.nn.Hardsigmoid()
         m.train(random())
@@ -377,14 +377,14 @@ class TestHardsigmoidModule(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_functional_hardsigmoid_with_random_data(test_case):
         device = random_device()
         x = random_tensor().to(device)
         y = torch.nn.functional.hardsigmoid(x, random_bool())
         return y
 
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_functional_hardsigmoid_with_0dim_data(test_case):
         device = random_device()
         x = random_tensor(ndim=0).to(device)
@@ -411,11 +411,11 @@ def do_test_softmax(batch_size: int, log_softmax: bool = False):
 
 @flow.unittest.skip_unless_1n1d()
 class TestSoftmax(flow.unittest.TestCase):
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_softmax_module_with_random_data(test_case):
         return do_test_softmax(batch_size=-1, log_softmax=False)
 
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_softmax_module_with_batch_size_equal_1024(test_case):
         return do_test_softmax(batch_size=1024, log_softmax=False)
 
@@ -435,11 +435,11 @@ class TestSoftmax(flow.unittest.TestCase):
 
 @flow.unittest.skip_unless_1n1d()
 class TestLogSoftmaxModule(flow.unittest.TestCase):
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_logsoftmax_module_with_random_data(test_case):
         return do_test_softmax(batch_size=-1, log_softmax=True)
 
-    @autotest()
+    @autotest(n=5)
     def test_softmax_module_with_batch_size_equal_1024(test_case):
         return do_test_softmax(batch_size=1024, log_softmax=True)
 
@@ -454,7 +454,7 @@ class TestLogSoftmaxModule(flow.unittest.TestCase):
 
 @flow.unittest.skip_unless_1n1d()
 class TestLogSigmoidModule(flow.unittest.TestCase):
-    @autotest()
+    @autotest(n=5)
     def test_logsigmoid_module_with_random_data(test_case):
         m = torch.nn.LogSigmoid()
         m.train(random())
@@ -464,7 +464,7 @@ class TestLogSigmoidModule(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest()
+    @autotest(n=5)
     def test_logsigmoid_module_with_0dim_data(test_case):
         m = torch.nn.LogSigmoid()
         m.train(random())
@@ -537,7 +537,7 @@ class TestSoftplusModule(flow.unittest.TestCase):
             arg[0](test_case, *arg[1:])
 
     @unittest.skip("pytorch softplus backward has bug")
-    @autotest()
+    @autotest(n=5)
     def test_softplus_module_with_random_data(test_case):
         m = torch.nn.Softplus(beta=random() | nothing(), threshold=random() | nothing())
         m.train(random())
@@ -550,7 +550,7 @@ class TestSoftplusModule(flow.unittest.TestCase):
 
 @flow.unittest.skip_unless_1n1d()
 class TestHardswishModule(flow.unittest.TestCase):
-    @autotest()
+    @autotest(n=5)
     def test_hardswish_module_with_random_data(test_case):
         m = torch.nn.Hardswish()
         m.train(random())
@@ -560,7 +560,7 @@ class TestHardswishModule(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest()
+    @autotest(n=5)
     def test_hardswish_module_with_0dim_data(test_case):
         m = torch.nn.Hardswish()
         m.train(random())
@@ -573,7 +573,7 @@ class TestHardswishModule(flow.unittest.TestCase):
 
 @flow.unittest.skip_unless_1n1d()
 class TestHardtanhModule(flow.unittest.TestCase):
-    @autotest()
+    @autotest(n=5)
     def test_hardtanh_module_with_random_data(test_case):
         m = torch.nn.Hardtanh(
             min_val=random().to(float) | nothing(),
@@ -586,7 +586,7 @@ class TestHardtanhModule(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest()
+    @autotest(n=5)
     def test_hardtanh_module_with_0dim_data(test_case):
         m = torch.nn.Hardtanh(
             min_val=random().to(float) | nothing(),
@@ -602,7 +602,7 @@ class TestHardtanhModule(flow.unittest.TestCase):
 
 @flow.unittest.skip_unless_1n1d()
 class TestLeakyReLUModule(flow.unittest.TestCase):
-    @autotest()
+    @autotest(n=5)
     def test_leakyrelu_module_with_random_data(test_case):
         m = torch.nn.LeakyReLU(negative_slope=random() | nothing())
         m.train(random())
@@ -624,7 +624,7 @@ class TestLeakyReLUModule(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest()
+    @autotest(n=5)
     def test_leakyrelu_module_with_0dim_data(test_case):
         m = torch.nn.LeakyReLU(negative_slope=random() | nothing())
         m.train(random())
@@ -719,14 +719,14 @@ class TestSoftsignModule(flow.unittest.TestCase):
 
 @flow.unittest.skip_unless_1n1d()
 class TestReluFunction(flow.unittest.TestCase):
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_flow_relu_with_random_data(test_case):
         device = random_device()
         x = random_tensor(ndim=2, dim1=3).to(device)
         y = torch.relu(x)
         return y
 
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_flow_relu_with_0dim_data(test_case):
         device = random_device()
         x = random_tensor(ndim=0).to(device)
@@ -736,14 +736,14 @@ class TestReluFunction(flow.unittest.TestCase):
 
 @flow.unittest.skip_unless_1n1d()
 class TestRelu6Function(flow.unittest.TestCase):
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_flow_nn_functional_relu6_with_random_data(test_case):
         device = random_device()
         x = random_tensor(ndim=2, dim1=3).to(device)
         y = torch.nn.functional.relu6(x)
         return y
 
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_flow_nn_functional_relu6_with_0dim_data(test_case):
         device = random_device()
         x = random_tensor(ndim=0).to(device)
@@ -753,14 +753,14 @@ class TestRelu6Function(flow.unittest.TestCase):
 
 @flow.unittest.skip_unless_1n1d()
 class TestLogSigmoidFunction(flow.unittest.TestCase):
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_flow_nn_functional_logsigmoid_with_random_data(test_case):
         device = random_device()
         x = random_tensor(ndim=2, dim1=3).to(device)
         y = torch.nn.functional.logsigmoid(x)
         return y
 
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_flow_nn_functional_logsigmoid_with_0dim_data(test_case):
         device = random_device()
         x = random_tensor(ndim=0).to(device)
@@ -790,7 +790,7 @@ class TestHardshrinkModule(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest(auto_backward=False, check_graph=True)
+    @autotest(n=5, auto_backward=False)
     def test_hardshrink_module_with_0_size_data(test_case):
         m = torch.nn.Hardshrink(lambd=random() | nothing())
         m.train(random())
@@ -823,7 +823,7 @@ class TestSoftshrinkModule(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest(auto_backward=False, check_graph=True)
+    @autotest(n=5, auto_backward=False)
     def test_softshrink_module_with_0_size_data(test_case):
         m = torch.nn.Softshrink(alpha=random() | nothing())
         m.train(random())
@@ -860,7 +860,7 @@ class TestThresholdModule(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest(auto_backward=False, check_graph=True)
+    @autotest(n=5, auto_backward=False)
     def test_threshold_module_with_0_size_data(test_case):
         m = torch.nn.Threshold(
             threshold=random() | nothing(), value=random() | nothing()