… python/oneflow/test/modules/test_math_ops.py
… python/oneflow/test/tensor/test_tensor_part_1.py
… python/oneflow/test/tensor/test_tensor_part_2.py