- [x] 减少 test_math_ops.py 运行时间。127s->47s。
- [x] 减少 test_loss.py 运行时间。168s->29s。
- [x] 减少 test_activation.py 运行时间。204s->68s。
- [x] 减少 test_tensor_part1.py 运行时间。119s->53s。
- [x] 减少 test_tensor_part2.py 运行时间。63s->38s。

以上几个耗时最高的Eager Test执行时间总和可以优化到原来执行时间的 (127+168+204+119+63) /(47+29+68+53+38)  = 681/235 约等于3倍。