@BBuf 2022-06-27
> 那要不要把 autotest 的默认 n 调整到 10（无脑建议）

剩下的测试基本都是单个op的测试，可能作者有一些其它考虑，暂时就不改了。